PROF_SAFE24_PR_GITHUB/requirements.txt
Flask
gunicorn
python-dotenv
requests
reportlab
